change name from libisr to libdio
